refactor: agent knowledge pane
