+++ web/src/sections/knowledge/agent-knowledge/interfaces.ts
@@ -1,0 +1,20 @@
+import type { ValidSources } from "@/lib/types";
+import type { HierarchyNodeSearchSummary } from "@/lib/hierarchy/interfaces";
+import type { SearchDocWithContent } from "@/lib/search/interfaces";
+
+export type KnowledgeView =
+  | "main"
+  | "add"
+  | "document-sets"
+  | "sources"
+  | "recent";
+
+export interface KnowledgeSearchResults {
+  docs: SearchDocWithContent[];
+  nodes: HierarchyNodeSearchSummary[];
+}
+
+export interface KnowledgeNavState {
+  view: KnowledgeView;
+  activeSource: ValidSources | undefined;
+}
